refactor: tcp 서버코드, 부모클래스로 추출
